Add Switch

+++ Java/src/main/java/com/guildedrose/InMemoryRepository.java
@@ -1,32 +1,33 @@
 package com.guildedrose;
 
 import com.guildedrose.entities.Item;
-import com.guildedrose.items.AgedItem;
-import com.guildedrose.items.EventItem;
-import com.guildedrose.items.GenericItem;
-import com.guildedrose.items.LegendaryItem;
+import com.guildedrose.items.*;
 
 import java.io.*;
-import java.nio.file.Files;
 import java.util.ArrayList;
 
 public class InMemoryRepository implements IInventoryRepository {
     //fichier txt
-    FileWriter fileWriter = new FileWriter("inventory.txt");
+    File file;
+
 
     ArrayList<Item> items = new ArrayList<Item>();
 
     public InMemoryRepository() throws IOException {
-        items.add(new AgedItem("Aged Brie", 5, 5));
-        items.add(new LegendaryItem("Sulfuras", 10, 1));
-        items.add(new EventItem("Backstage passes", 10, 0));
-        items.add(new GenericItem("Item normal", 7, 8));
+        if (file.exists()) {
+            FileWriter fileWriter = new FileWriter("inventory.txt");
+            items.add(new AgedItem("Aged Brie", 5, 5));
+            items.add(new LegendaryItem("Sulfuras", 10, 1));
+            items.add(new EventItem("Backstage passes", 10, 0));
+            items.add(new GenericItem("Item normal", 7, 8));
 
-        SaveInventory(items);
+            SaveInventory(items);
+        }
     }
 
     @Override
     public ArrayList<Item> GetInventory() {
+        ArrayList<Item> items = new ArrayList<>();
         String line = "";
         String splitBy = ",";
         try {
@@ -34,7 +35,23 @@ public class InMemoryRepository implements IInventoryRepository {
             while ((line = bufferedReader.readLine()) != null)
             {
                 String[] item = line.split(splitBy);
-                System.out.println("Employee [Item=" + item[0] + ", Name=" + item[1] + ", SellIn=" + item[2] + ", Quality=" + item[3] + "]");
+                switch(item[0]) {
+                    case "AgedItem":
+                        items.add(new AgedItem(item[1], Integer.parseInt(item[2]), Integer.parseInt(item[3])));
+                        break;
+                    case "LegendaryItem":
+                        items.add(new LegendaryItem(item[1], Integer.parseInt(item[2]), Integer.parseInt(item[3])));
+                        break;
+                    case "EventItem":
+                        items.add(new EventItem(item[1], Integer.parseInt(item[2]), Integer.parseInt(item[3])));
+                        break;
+                    case "GenericItem":
+                        items.add(new GenericItem(item[1], Integer.parseInt(item[2]), Integer.parseInt(item[3])));
+                        break;
+                    case "ConjuredItem":
+                        items.add(new ConjuredItem(item[1], Integer.parseInt(item[2]), Integer.parseInt(item[3])));
+                        break;
+                }
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -45,11 +62,10 @@ public class InMemoryRepository implements IInventoryRepository {
     @Override
     public void SaveInventory(ArrayList<Item> items) {
         try {
-            FileWriter fileWriter = new FileWriter("inventory.txt");
-            BufferedWriter bw = new BufferedWriter(fileWriter);
+            BufferedWriter bw = new BufferedWriter(new FileWriter("inventory.txt"));
             for (Item item : items) {
                 bw.write(String.format("%s,%s,%d,%d", item.getClass().getSimpleName(), item.getNom(), item.getSellin(), item.getQuality()));
-                bw.write("\n");
+                bw.newLine();
             }
             bw.close();
         } catch (Exception e) {
